CHISL-31 Final server refactoring

+++ chislius_server/src/main/java/ru/hse/chislius_server/service/RoomService.java
@@ -4,6 +4,7 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
+import ru.hse.chislius_server.dto.update.MessageType;
 import ru.hse.chislius_server.exception.DataValidationException;
 import ru.hse.chislius_server.exception.EntityNotFoundException;
 import ru.hse.chislius_server.exception.GenerationTimeoutException;
@@ -14,6 +15,7 @@ import ru.hse.chislius_server.model.room.Room;
 import ru.hse.chislius_server.model.room.RoomState;
 import ru.hse.chislius_server.model.room.RoomType;
 import ru.hse.chislius_server.repository.RoomRepository;
+import ru.hse.chislius_server.repository.UserRepository;
 
 import java.util.Collections;
 import java.util.Random;
@@ -25,6 +27,7 @@ public class RoomService {
     private final RoomRepository roomRepository;
     private final GameService gameService;
     private final UpdateService updateService;
+    private final UserRepository userRepository;
 
     @Value("${config.room.capacity.public}")
     private int PUBLIC_ROOM_CAPACITY;
@@ -34,18 +37,6 @@ public class RoomService {
 
     @Value("${config.room.capacity.private.max}")
     private int PRIVATE_ROOM_CAPACITY_MAX;
-
-    public void validateUserNotInRoom(long userId) {
-        Room room = roomRepository.findRoomByUserId(userId).orElse(null);
-        if (room != null) {
-            throw new DataValidationException("Вы должны выйти из комнаты, чтобы совершить это дейcтвие");
-        }
-    }
-
-    public Room getCurrentRoom(long userId) {
-        return roomRepository.findRoomByUserId(userId)
-                .orElseThrow(() -> new DataValidationException("Вы должны войти в комнату, чтобы совершить это дейcтвие"));
-    }
 
     public Room createPrivateRoom(long userId, int capacity, int gameMode) {
         validateUserNotInRoom(userId);
@@ -90,12 +81,13 @@ public class RoomService {
     }
 
     public void leaveRoom(long userId) {
-        Room room = getCurrentRoom(userId);
-        if (!room.getUserIds().remove(userId)) {
-            throw new DataValidationException("Не удалось покинуть комнату");
+        Room room = roomRepository.findRoomByUserId(userId).orElse(null);
+        if (room == null || !room.getUserIds().remove(userId)) {
+            return;
         }
-        updateService.sendUpdate(userId, room);
         roomRepository.deleteUserIdFromIndex(userId);
+        updateService.sendUpdate(userId);
+        updateService.sendMessage(userId, MessageType.LEAVE, userId);
         if (room.getUserIds().isEmpty()) {
             deleteRoom(room);
             return;
@@ -105,16 +97,62 @@ public class RoomService {
         } else {
             room.getPlayerIds().remove(userId);
         }
-        broadcastRoom(room);
+        if (room.getState() != RoomState.FINISH) {
+            notifyRoom(room, MessageType.LEAVE, userId);
+            broadcastRoom(room);
+        }
     }
 
-    private void deleteRoom(Room room) {
-        roomRepository.deleteRoom(room);
-        room.setState(RoomState.DELETE);
+    public void flipCard(long userId, int id) {
+        Room room = getCurrentRoom(userId);
+        if (room.getCurrentPlayerId() == userId) {
+            if (!gameService.openCard(room.getGame(), id - 1)) {
+                throw new DataValidationException("Вы не можете перевернуть эту карту");
+            }
+            broadcastRoom(room);
+        } else {
+            throw new DataValidationException("Сейчас не ваш ход");
+        }
     }
 
-    public void broadcastRoom(Room room) {
-        room.getUserIds().forEach((userId) -> updateService.sendUpdate(userId, room));
+    public void submitMove(long userId) {
+        Room room = getCurrentRoom(userId);
+        if (room.getCurrentPlayerId() == userId) {
+            int d = gameService.doMove(room.getGame());
+            if (d == 0) {
+                throw new DataValidationException("Требуется выбрать минимум 2 карты");
+            }
+            room.getScores().put(userId, room.getScores().get(userId) + d);
+            room.getPlayerIds().add(0, userId);
+            room.setCurrentPlayerId(room.getPlayerIds().removeLast());
+            checkGameOver(room);
+            broadcastRoom(room);
+            if (d > 0) {
+                notifyRoom(room, MessageType.SUCCESS, userId, d);
+            } else {
+                notifyRoom(room, MessageType.FAIL, userId, -d);
+            }
+        } else {
+            throw new DataValidationException("Сейчас не ваш ход");
+        }
+    }
+
+    public void skipMove(long userId) {
+        Room room = getCurrentRoom(userId);
+        if (room.getCurrentPlayerId() == userId) {
+            gameService.skipMove(room.getGame());
+            room.getPlayerIds().add(0, userId);
+            room.setCurrentPlayerId(room.getPlayerIds().removeLast());
+            broadcastRoom(room);
+            notifyRoom(room, MessageType.SKIP, userId);
+        } else {
+            throw new DataValidationException("Сейчас не ваш ход");
+        }
+    }
+
+    private Room getCurrentRoom(long userId) {
+        return roomRepository.findRoomByUserId(userId)
+                .orElseThrow(() -> new DataValidationException("Вы должны войти в комнату, чтобы совершить это дейcтвие"));
     }
 
     private Room getByCode(String code) {
@@ -137,12 +175,6 @@ public class RoomService {
         throw new GenerationTimeoutException();
     }
 
-    private String generateRoomCode() {
-        Random random = new Random();
-        int number = random.nextInt(1000000);
-        return String.format("%06d", number);
-    }
-
     private void joinRoom(long userId, Room room) {
         if (room.getState() != RoomState.WAIT) {
             throw new DataValidationException("В комнате уже началась игра");
@@ -151,10 +183,41 @@ public class RoomService {
             throw new DataValidationException("В комнате нет свободных мест");
         }
         room.getUserIds().add(userId);
+        notifyRoom(room, MessageType.JOIN, userId);
         if (room.getUserIds().size() == room.getCapacity()) {
             startGame(room);
         }
         roomRepository.createUserIdIndex(userId, room);
+    }
+
+    private void deleteRoom(Room room) {
+        roomRepository.deleteRoom(room);
+        room.setState(RoomState.DELETE);
+    }
+
+    public void broadcastRoom(Room room) {
+        room.getUserIds().forEach((userId) -> updateService.sendUpdate(userId, room));
+    }
+
+    private void notifyRoom(Room room, MessageType type, long userId) {
+        room.getUserIds().forEach((destinationId) -> updateService.sendMessage(destinationId, type, userId));
+    }
+
+    private void notifyRoom(Room room, MessageType type, long userId, int score) {
+        room.getUserIds().forEach((destinationId) -> updateService.sendMessage(destinationId, type, userId, score));
+    }
+
+    private String generateRoomCode() {
+        Random random = new Random();
+        int number = random.nextInt(1000000);
+        return String.format("%06d", number);
+    }
+
+    private void validateUserNotInRoom(long userId) {
+        Room room = roomRepository.findRoomByUserId(userId).orElse(null);
+        if (room != null) {
+            throw new DataValidationException("Вы должны выйти из комнаты, чтобы совершить это дейcтвие");
+        }
     }
 
     private void startGame(Room room) {
@@ -165,51 +228,20 @@ public class RoomService {
         Collections.shuffle(room.getPlayerIds());
         room.setCurrentPlayerId(room.getPlayerIds().removeLast());
         room.getUserIds().forEach((userId) -> room.getScores().put(userId, 0));
+        checkGameOver(room);
         broadcastRoom(room);
-    }
-
-    public void flipCard(long userId, int id) {
-        Room room = getCurrentRoom(userId);
-        if (room.getCurrentPlayerId() == userId) {
-            if (!gameService.openCard(room.getGame(), id - 1)) {
-                throw new DataValidationException("Вы не можете перевернуть эту карту");
-            }
-            broadcastRoom(room);
-        } else {
-            throw new DataValidationException("Сейчас не ваш ход");
-        }
-    }
-
-    public void skipMove(long userId) {
-        Room room = getCurrentRoom(userId);
-        if (room.getCurrentPlayerId() == userId) {
-            gameService.skipMove(room.getGame());
-            room.getPlayerIds().add(0, userId);
-            room.setCurrentPlayerId(room.getPlayerIds().removeLast());
-            checkGameOver(room);
-            broadcastRoom(room);
-        } else {
-            throw new DataValidationException("Сейчас не ваш ход");
-        }
-    }
-
-    public void submitMove(long userId) {
-        Room room = getCurrentRoom(userId);
-        if (room.getCurrentPlayerId() == userId) {
-            int d = gameService.doMove(room.getGame());
-            room.getScores().put(userId, room.getScores().get(userId) + d);
-            room.getPlayerIds().add(0, userId);
-            room.setCurrentPlayerId(room.getPlayerIds().removeLast());
-            checkGameOver(room);
-            broadcastRoom(room);
-        } else {
-            throw new DataValidationException("Сейчас не ваш ход");
-        }
     }
 
     private void checkGameOver(Room room) {
         if (!gameService.canMove(room.getGame())) {
             room.setState(RoomState.FINISH);
+            room.getScores().forEach((userId, score) -> userRepository.findById(userId).ifPresent(user -> {
+                user.setRating(user.getRating() + score);
+                if (user.getRating() < 0) {
+                    user.setRating(0);
+                }
+                userRepository.save(user);
+            }));
         }
     }
 }
